move open into model
