+++ frontend/src/scenes/notebooks/Notebook/utils.ts
@@ -1,27 +1,20 @@
 // Helpers for Kea issue with double importing
 import {
-    JSONContent as TTJSONContent,
-    Editor as TTEditor,
     ChainedCommands as EditorCommands,
+    Editor as TTEditor,
     FocusPosition as EditorFocusPosition,
-    Range as EditorRange,
     getText,
+    JSONContent as TTJSONContent,
+    Range as EditorRange,
 } from '@tiptap/core'
 import { Node as PMNode } from '@tiptap/pm/model'
 import { NodeViewProps } from '@tiptap/react'
-import { NotebookNodeType, NotebookTarget } from '~/types'
-import { notebookLogicType } from 'scenes/notebooks/Notebook/notebookLogicType'
-import { notebookPopoverLogic } from 'scenes/notebooks/Notebook/notebookPopoverLogic'
-import { urls } from 'scenes/urls'
-import { notebookLogic } from 'scenes/notebooks/Notebook/notebookLogic'
-import { BuiltLogic } from 'kea'
-import { router } from 'kea-router'
+import { NotebookNodeType } from '~/types'
 
 /* eslint-disable @typescript-eslint/no-empty-interface */
 export interface Node extends PMNode {}
 export interface JSONContent extends TTJSONContent {}
 /* eslint-enable @typescript-eslint/no-empty-interface */
-// export type FocusPosition = number | boolean | 'start' | 'end' | 'all' | null
 
 export {
     ChainedCommands as EditorCommands,
@@ -94,35 +87,4 @@ export function defaultNotebookContent(title?: string, content?: JSONContent[]):
     }
 
     return { type: 'doc', content: initialContent }
-}
-
-export const openNotebook = async (
-    notebookId: string,
-    target: NotebookTarget = NotebookTarget.Auto,
-    focus: EditorFocusPosition = null,
-    // operations to run against the notebook once it has opened and the editor is ready
-    onOpen: (logic: BuiltLogic<notebookLogicType>) => void = () => {}
-): Promise<void> => {
-    const popoverLogic = notebookPopoverLogic.findMounted()
-
-    if (NotebookTarget.Popover === target) {
-        popoverLogic?.actions.setVisibility('visible')
-    }
-
-    if (popoverLogic?.values.visibility === 'visible') {
-        popoverLogic?.actions.selectNotebook(notebookId)
-    } else {
-        router.actions.push(urls.notebookEdit(notebookId))
-    }
-
-    popoverLogic?.actions.setInitialAutofocus(focus)
-
-    const theNotebookLogic = notebookLogic({ shortId: notebookId })
-    const unmount = theNotebookLogic.mount()
-
-    try {
-        onOpen(theNotebookLogic)
-    } finally {
-        unmount()
-    }
 }
